Implementar funcionalidade salvar produto

+++ src/ProdutosDTO.java
@@ -1,3 +1,6 @@
+
+import java.text.DecimalFormat;
+
 /*
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
@@ -10,7 +13,7 @@
 public class ProdutosDTO {
     private Integer id;
     private String nome;
-    private Integer valor;
+    private float valor;
     private String status;
 
     public Integer getId() {
@@ -21,7 +24,7 @@ public class ProdutosDTO {
         return nome;
     }
 
-    public Integer getValor() {
+    public float getValor() {
         return valor;
     }
 
@@ -37,7 +40,7 @@ public class ProdutosDTO {
         this.nome = nome;
     }
 
-    public void setValor(Integer valor) {
+    public void setValor(float valor) {
         this.valor = valor;
     }
